// jems-ui/src/app/call/notifications-settings/notification-variable.enum.ts
export enum NotificationVariable {
  PROGRAMME_NAME = "PROGRAMME_NAME",
  CALL_ID = "CALL_ID",
  CALL_NAME = "CALL_NAME",
  PROJECT_ID = "PROJECT_ID",
  PROJECT_IDENTIFIER = "PROJECT_IDENTIFIER",
  PROJECT_ACRONYM = "PROJECT_ACRONYM",
  PARTNER_ID = "PARTNER_ID",
  PARTNER_ROLE = "PARTNER_ROLE",
  PARTNER_NUMBER = "PARTNER_NUMBER",
  PARTNER_REPORT_ID = "PARTNER_REPORT_ID",
  PARTNER_REPORT_NUMBER = "PARTNER_REPORT_NUMBER",
  PROJECT_REPORT_ID = "PROJECT_REPORT_ID",
  PROJECT_REPORT_NUMBER = "PROJECT_REPORT_NUMBER"
}